+++ carp_backend/lib/carp_backend.g.dart
@@ -8,19 +8,20 @@ part of carp_backend;
 
 CarpDataEndPoint _$CarpDataEndPointFromJson(Map<String, dynamic> json) {
   return CarpDataEndPoint(
-      _$enumDecodeNullable(_$CarpUploadMethodEnumMap, json['upload_method']),
-      name: json['name'] as String,
-      uri: json['uri'] as String,
-      clientId: json['client_id'] as String,
-      clientSecret: json['client_secret'] as String,
-      email: json['email'] as String,
-      password: json['password'] as String,
-      collection: json['collection'] as String,
-      deleteWhenUploaded: json['delete_when_uploaded'] as bool,
-      bufferSize: json['buffer_size'],
-      zip: json['zip'],
-      encrypt: json['encrypt'],
-      publicKey: json['public_key'])
+    _$enumDecodeNullable(_$CarpUploadMethodEnumMap, json['upload_method']),
+    name: json['name'] as String,
+    uri: json['uri'] as String,
+    clientId: json['client_id'] as String,
+    clientSecret: json['client_secret'] as String,
+    email: json['email'] as String,
+    password: json['password'] as String,
+    collection: json['collection'] as String,
+    deleteWhenUploaded: json['delete_when_uploaded'] as bool,
+    bufferSize: json['buffer_size'],
+    zip: json['zip'],
+    encrypt: json['encrypt'],
+    publicKey: json['public_key'],
+  )
     ..c__ = json['c__'] as String
     ..type = json['type'] as String;
 }
@@ -53,29 +54,41 @@ Map<String, dynamic> _$CarpDataEndPointToJson(CarpDataEndPoint instance) {
   return val;
 }
 
-T _$enumDecode<T>(Map<T, dynamic> enumValues, dynamic source) {
+T _$enumDecode<T>(
+  Map<T, dynamic> enumValues,
+  dynamic source, {
+  T unknownValue,
+}) {
   if (source == null) {
     throw ArgumentError('A value must be provided. Supported values: '
         '${enumValues.values.join(', ')}');
   }
-  return enumValues.entries
-      .singleWhere((e) => e.value == source,
-          orElse: () => throw ArgumentError(
-              '`$source` is not one of the supported values: '
-              '${enumValues.values.join(', ')}'))
-      .key;
+
+  final value = enumValues.entries
+      .singleWhere((e) => e.value == source, orElse: () => null)
+      ?.key;
+
+  if (value == null && unknownValue == null) {
+    throw ArgumentError('`$source` is not one of the supported values: '
+        '${enumValues.values.join(', ')}');
+  }
+  return value ?? unknownValue;
 }
 
-T _$enumDecodeNullable<T>(Map<T, dynamic> enumValues, dynamic source) {
+T _$enumDecodeNullable<T>(
+  Map<T, dynamic> enumValues,
+  dynamic source, {
+  T unknownValue,
+}) {
   if (source == null) {
     return null;
   }
-  return _$enumDecode<T>(enumValues, source);
+  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
 }
 
-const _$CarpUploadMethodEnumMap = <CarpUploadMethod, dynamic>{
+const _$CarpUploadMethodEnumMap = {
   CarpUploadMethod.DATA_POINT: 'DATA_POINT',
   CarpUploadMethod.BATCH_DATA_POINT: 'BATCH_DATA_POINT',
   CarpUploadMethod.FILE: 'FILE',
-  CarpUploadMethod.DOCUMENT: 'DOCUMENT'
+  CarpUploadMethod.DOCUMENT: 'DOCUMENT',
 };