+++ carp_esense_package/lib/esense.g.dart
@@ -8,13 +8,14 @@ part of esense;
 
 ESenseMeasure _$ESenseMeasureFromJson(Map<String, dynamic> json) {
   return ESenseMeasure(
-      json['type'] == null
-          ? null
-          : MeasureType.fromJson(json['type'] as Map<String, dynamic>),
-      name: json['name'],
-      enabled: json['enabled'],
-      deviceName: json['device_name'] as String,
-      samplingRate: json['sampling_rate'] as int)
+    json['type'] == null
+        ? null
+        : MeasureType.fromJson(json['type'] as Map<String, dynamic>),
+    name: json['name'],
+    enabled: json['enabled'],
+    deviceName: json['device_name'] as String,
+    samplingRate: json['sampling_rate'] as int,
+  )
     ..c__ = json['c__'] as String
     ..configuration = (json['configuration'] as Map<String, dynamic>)?.map(
       (k, e) => MapEntry(k, e as String),
@@ -42,8 +43,9 @@ Map<String, dynamic> _$ESenseMeasureToJson(ESenseMeasure instance) {
 
 ESenseButtonDatum _$ESenseButtonDatumFromJson(Map<String, dynamic> json) {
   return ESenseButtonDatum(
-      deviceName: json['device_name'] as String,
-      pressed: json['pressed'] as bool)
+    deviceName: json['device_name'] as String,
+    pressed: json['pressed'] as bool,
+  )
     ..id = json['id'] as String
     ..timestamp = json['timestamp'] == null
         ? null
@@ -68,14 +70,14 @@ Map<String, dynamic> _$ESenseButtonDatumToJson(ESenseButtonDatum instance) {
 
 ESenseSensorDatum _$ESenseSensorDatumFromJson(Map<String, dynamic> json) {
   return ESenseSensorDatum(
-      deviceName: json['device_name'] as String,
-      timestamp: json['timestamp'] == null
-          ? null
-          : DateTime.parse(json['timestamp'] as String),
-      packetIndex: json['packet_index'] as int,
-      accel: (json['accel'] as List)?.map((e) => e as int)?.toList(),
-      gyro: (json['gyro'] as List)?.map((e) => e as int)?.toList())
-    ..id = json['id'] as String;
+    deviceName: json['device_name'] as String,
+    timestamp: json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String),
+    packetIndex: json['packet_index'] as int,
+    accel: (json['accel'] as List)?.map((e) => e as int)?.toList(),
+    gyro: (json['gyro'] as List)?.map((e) => e as int)?.toList(),
+  )..id = json['id'] as String;
 }
 
 Map<String, dynamic> _$ESenseSensorDatumToJson(ESenseSensorDatum instance) {